Add direct navigation to panes.
This adds a simple routing system to the IDE / Project page that allows direct
linking to a file or resource, as well as to the settings, compilation, or
github panes. This is acclomplished by using the window.location.hash
property, and removes some of the burden from the sidebar's javascript
for managing the current view state.